Eliminate redundant clone if tnt feature is off
THe way this code was organized caused the last store constructor to
have a redundant clone clippy warning when track and trace isnt being
used because the connection pool wouldnt be used after the clone. This
just moves the track and trace store declaration to the top to get rid
of that error.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>